+++ kwimage/im_demodata.py
@@ -581,7 +581,7 @@ grab_test_image_fpath.keys = lambda: _TEST_IMAGES.keys()
 
 
 def checkerboard(num_squares='auto', square_shape='auto', dsize=(512, 512),
-                 dtype=float, on_value=1, off_value=0):
+                 dtype=float, on_value=1, off_value=0, bayer_value=None):
     """
     Creates a checkerboard image, mainly for use in testing.
 
@@ -605,10 +605,21 @@ def checkerboard(num_squares='auto', square_shape='auto', dsize=(512, 512),
             The value off the other checker. Defaults to 0.
             Can also be the name of a color.
 
+        bayer_value (Number | int | str | None):
+            If specified, adds a third value to the checkerboard similar to a
+            Bayer pattern [WikiBayerFilter]_. The on and off values become the
+            value for the one quater parts of the bayer pattern and this is the
+            value for the remaining half.
+
+    Returns:
+        ndarray: a numpy array representing a checkerboard pattern
+
     References:
         .. [SO2169478] https://stackoverflow.com/questions/2169478/how-to-make-a-checkerboard-in-numpy
+        .. [WikiBayerFilter] https://en.wikipedia.org/wiki/Bayer_filter
 
     Example:
+        >>> # Various invocations of checkerboard
         >>> import kwimage
         >>> import numpy as np
         >>> img = kwimage.checkerboard()
@@ -618,16 +629,35 @@ def checkerboard(num_squares='auto', square_shape='auto', dsize=(512, 512),
         >>> print(kwimage.checkerboard(square_shape=3, dsize=(1451, 1163)).shape)
         >>> print(kwimage.checkerboard(square_shape=3, dsize=(1202, 956)).shape)
         >>> print(kwimage.checkerboard(dsize=(4, 4), on_value=(255, 0, 0), off_value=(0, 0, 1), dtype=np.uint8))
+        >>> print(kwimage.checkerboard(dsize=(4, 4), on_value=(255, 0, 0), off_value=(0, 0, 1), bayer_value=(1, 9, 1), dtype=np.uint8))
 
     Example:
         >>> import kwimage
-        >>> img = kwimage.checkerboard(
+        >>> img1c = kwimage.checkerboard(dsize=(64, 64))
+        >>> img3c = kwimage.checkerboard(
         >>>     dsize=(64, 64), on_value='kw_green', off_value='kw_blue')
         >>> # xdoctest: +REQUIRES(--show)
         >>> # xdoctest: +REQUIRES(module:kwplot)
         >>> import kwplot
         >>> kwplot.autoplt()
-        >>> kwplot.imshow(img)
+        >>> kwplot.figure().clf()
+        >>> kwplot.imshow(img1c, pnum=(1, 2, 1), title='1 Channel Basic Checkerboard')
+        >>> kwplot.imshow(img3c, pnum=(1, 2, 2), title='3 Channel Basic Checkerboard')
+        >>> kwplot.show_if_requested()
+
+    Example:
+        >>> import kwimage
+        >>> img1c = kwimage.checkerboard(bayer_value=0.5)
+        >>> img3c = kwimage.checkerboard(
+        >>>     on_value='kw_blue', off_value='kw_red',
+        >>>     bayer_value='kw_green')
+        >>> # xdoctest: +REQUIRES(--show)
+        >>> # xdoctest: +REQUIRES(module:kwplot)
+        >>> import kwplot
+        >>> kwplot.autoplt()
+        >>> kwplot.figure().clf()
+        >>> kwplot.imshow(img1c, pnum=(1, 2, 1), title='1 Channel Bayer Checkerboard')
+        >>> kwplot.imshow(img3c, pnum=(1, 2, 2), title='3 Channel Bayer Checkerboard')
         >>> kwplot.show_if_requested()
 
     Ignore:
@@ -635,12 +665,51 @@ def checkerboard(num_squares='auto', square_shape='auto', dsize=(512, 512),
         globals().update(xdev.get_func_kwargs(kwimage.checkerboard))
     """
     import numpy as np
-    import kwimage
-    if num_squares == 'auto' and square_shape == 'auto':
-        num_squares = 8
 
     want_w, want_h = dsize
 
+    # Resolve number of pixels for the image and the square
+    h, w, num_h, num_w = _resolve_checkerboard_shape_args(square_shape,
+                                                          num_squares, want_w,
+                                                          want_h)
+
+    # Resolve the color values
+    on_value = _resolve_checkerboard_color_arg(on_value, dtype)
+    off_value = _resolve_checkerboard_color_arg(off_value, dtype)
+    if bayer_value is not None:
+        bayer_value = _resolve_checkerboard_color_arg(bayer_value, dtype)
+        # For efficiency swap the bayer value with the off value in the
+        # following logic.
+        bayer_value, off_value =  off_value, bayer_value
+
+    # All paramters have been resolved, build the image.
+    num_pairs_w = int(num_w // 2)
+    num_pairs_h = int(num_h // 2)
+    base = np.array([
+        [on_value, off_value] * num_pairs_w,
+        [off_value, on_value] * num_pairs_w
+    ] * num_pairs_h, dtype=dtype)
+
+    if len(base.shape) == 3:
+        base = base.transpose([2, 0, 1])
+
+    if bayer_value is not None:
+        base[..., 1::2, 1::2] = np.array(bayer_value)[..., None, None]
+
+    expansion = np.ones((h, w), dtype=dtype)
+    img = np.kron(base, expansion)[0:want_h, 0:want_w]
+    if len(base.shape) == 3:
+        img = img.transpose([1, 2, 0])
+    return img
+
+
+def _resolve_checkerboard_shape_args(square_shape, num_squares, want_w,
+                                     want_h):
+
+    if num_squares == 'auto' and square_shape == 'auto':
+        num_squares = 8
+
+    # Resolve the pixel width and height of each square.
     if square_shape != 'auto':
         if not ub.iterable(square_shape):
             square_shape = [square_shape, square_shape]
@@ -649,6 +718,7 @@ def checkerboard(num_squares='auto', square_shape='auto', dsize=(512, 512),
     else:
         gen_h, gen_w = _next_multiple_of(want_h, 4), _next_multiple_of(want_w, 4)
 
+    # Resolve the number of squares in each row and column.
     if num_squares == 'auto':
         assert square_shape != 'auto'
         if not ub.iterable(square_shape):
@@ -673,26 +743,14 @@ def checkerboard(num_squares='auto', square_shape='auto', dsize=(512, 512),
 
     num_h, num_w = num_squares
 
-    if isinstance(on_value, str):
-        on_value = kwimage.Color(on_value).forimage(dtype)
+    return h, w, num_h, num_w
 
-    if isinstance(off_value, str):
-        off_value = kwimage.Color(off_value).forimage(dtype)
 
-    num_pairs_w = int(num_w // 2)
-    num_pairs_h = int(num_h // 2)
-    base = np.array([
-        [on_value, off_value] * num_pairs_w,
-        [off_value, on_value] * num_pairs_w
-    ] * num_pairs_h, dtype=dtype)
-
-    if len(base.shape) == 3:
-        base = base.transpose([2, 0, 1])
-    expansion = np.ones((h, w), dtype=dtype)
-    img = np.kron(base, expansion)[0:want_h, 0:want_w]
-    if len(base.shape) == 3:
-        img = img.transpose([1, 2, 0])
-    return img
+def _resolve_checkerboard_color_arg(value, dtype):
+    import kwimage
+    if isinstance(value, str):
+        value = kwimage.Color(value).forimage(dtype)
+    return value
 
 
 def _next_power_of_two(x):